Work around the fact that sr-led-redgreen-0805 identifies different parts in the upstream power board and motor board BOMs
Strictly, the one on the power board should be called sr-led-redgreen-dual,
as they contain two independent LEDs and hence have four pins rather than two.
A true sr-led-redgreen-0805, as used on the motor board, contains two
LEDs connected in opposite directions in parallel (such that one is lit
when a forward voltage is applied and the other is lit when a reverse
voltage is applied); as such this package only has two pins.